backend/requirements.txt
blis==0.2.4
certifi==2019.6.16
chardet==3.0.4
cymem==2.0.2
#de-core-news-sm==2.1.0
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.9.4
fuzzywuzzy==0.17.0
idna==2.8
lda==1.1.0
murmurhash==1.0.2
mysqlclient==1.4.2.post1
nltk==3.4.4
numpy==1.16.4
pandas==0.25.0
pbr==3.1.1
plac==0.9.6
preshed==2.0.1
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.1
requests==2.22.0
six==1.12.0
snowballstemmer==1.2.0
spacy==2.1.6
sqlparse==0.3.0
srsly==0.0.7
thinc==7.0.8
tqdm==4.32.2
urllib3==1.25.3
wasabi==0.2.2
